Fix sign extended operator warning.

+++ ICSharpCode.NRefactory/Completion/FrameworkLookup.cs
@@ -407,7 +407,8 @@ namespace ICSharpCode.NRefactory.Completion
 			Dictionary<ulong, ushort> frameworkLookupTable = new Dictionary<ulong, ushort> ();
 			ushort GetLookup (string packageName, string assemblyName, string ns)
 			{
-				var id = unchecked ((ulong)ns.GetHashCode () << 32 | (ulong)assemblyName.GetHashCode ());
+				var id = unchecked ((ulong)ns.GetHashCode () << 32 | 
+				                    (ulong)(uint)assemblyName.GetHashCode ());
 				ushort value;
 				if (frameworkLookupTable.TryGetValue (id, out value))
 					return value;
